move Interface to Contract namespace
