+++ tests/Event/TestOutputListener.php
@@ -4,8 +4,8 @@ declare(strict_types=1);
 
 namespace GitWrapper\Test\Event;
 
+use GitWrapper\Contract\Event\GitOutputListenerInterface;
 use GitWrapper\Event\GitOutputEvent;
-use GitWrapper\Event\GitOutputListenerInterface;
 
 final class TestOutputListener implements GitOutputListenerInterface
 {
